+++ src/App.js
@@ -2,7 +2,8 @@ import { Routes, Route, NavLink } from "react-router-dom";
 import Home from "./Home/Home";
 import MyRouter from "./MyRouter/MyRouter";
 import styles from "./App.module.css";
-import { Cars } from "./Cars/Cars";
+import { Cars } from "./Cars";
+
